requirements.txt
Django==3.2.4
django-rest-framework==3.13.1